Wait for responses indefinitely
Say if we had 1 networking interface, shout_for_peers() used to send 1
broadcasting request and wait exactly for one response. Then it would
terminate the stream.

Now it will send 1 broadcast request and wait for responses infinitely
until the stream is cancelled explicitly. This is a good thing, cause
say we are broadcasting peer discovery request to 192.168.1.255 and
receive 1 response say from 192.168.1.100. Then the stream will yield
one item and terminate.  This is not ok, cause say 192.168.1.200 could
also have replied to the same broadcasted message. Now this won't happen
and we will collect all responses.